Fix time zone issues in epgxmltv
The start and stop timestamps were previously formatted/specified in
UTC. This was incorrect, because the offset of the local timezone is
appended to these timestamps, causing consumers to interpret the
formatted UTC timestamps in another time zone instead.

Timestamps are now put out in the local time zone, matching the appended
time zone offset.